funcionalidades basicas testadas

See merge request escoladeti2022time03/via-lactea-api!48

+++ src/main/java/rocketshell/vialactea/domain/Animal.java
@@ -1,15 +1,8 @@
 package rocketshell.vialactea.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Date;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -18,6 +11,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import rocketshell.vialactea.base.BaseEntity;
+
+import javax.persistence.*;
 
 
 @Entity
@@ -33,7 +28,7 @@ public class Animal extends BaseEntity{
    private String especie;
    
    
-   private Float peso;
+   private BigDecimal peso;
    
    private String raca;
